% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm.R
\name{as.data_model.data.frame}
\alias{as.data_model.data.frame}
\title{Coerce a data frame to a data model}
\usage{
\method{as.data_model}{data.frame}(x)
}
\description{
Coerce a data frame to a data model
}
\keyword{internal}
