% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse_db_model.R
\name{browse_db_model}
\alias{browse_db_model}
\title{Make a database diagram using a connection and display to browser}
\usage{
browse_db_model(connection = NULL, dbType = "sqlserver",
  rankdir = "BT", view_type = "keys_only", focus = NULL,
  wd = getwd(), fileName = "dbDiagram.svg")
}
\arguments{
\item{connection}{A database connection.}

\item{dbType}{One of 'sqlserver', or 'postgres'.}

\item{rankdir}{Direction of diagram default is 'BT'.}

\item{view_type}{One of 'all', 'keys_only', or 'view_only'.}

\item{focus}{A character vector of tables to be filtered to.}

\item{wd}{The working directory to use.}

\item{fileName}{The fileName.}
}
\description{
Make a database diagram using a connection and display to browser
}
\examples{
browse_db_model()
}
\keyword{database}
\keyword{diagram}
