% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm.R
\name{dm_add_reference_}
\alias{dm_add_reference_}
\title{Add reference}
\usage{
dm_add_reference_(dm, table, column, ref = NULL, ref_col = NULL)
}
\arguments{
\item{dm}{A data model object}

\item{table}{Table name}

\item{column}{Column(s) name}

\item{ref}{Referenced table name}

\item{ref_col}{Referenced column(s) name}
}
\value{
New data model object
}
\description{
Adds reference to existing data model object
}
