% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm.R
\name{dm_add_references}
\alias{dm_add_references}
\title{Add references}
\usage{
dm_add_references(dm, ...)
}
\arguments{
\item{dm}{Data model object}

\item{...}{Logical expressions in format table1$column1 == table2$column2}
}
\description{
Adds references defined with logical expressions from data frames
  in format table1$column1 == table2$column2
}
