% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/graph.R
\name{dm_create_graph}
\alias{dm_create_graph}
\title{Create a graph object from data model object}
\usage{
dm_create_graph(dm, rankdir = "BT", graph_name = "Data Model",
  graph_attrs = "", node_attrs = "", edge_attrs = "", view_type = "all",
  focus = NULL, col_attr = "column")
}
\arguments{
\item{dm}{A data model object}

\item{rankdir}{Graph attribute for direction (eg. 'BT' = bottom --> top)}

\item{graph_name}{A graph name}

\item{graph_attrs}{Additional graph attributes}

\item{node_attrs}{Additional node attributes}

\item{edge_attrs}{Additional edge attributes}

\item{view_type}{Can be "all" (by default), "keys_only" or "title_only".
It defines the level of details for the table rendering
(all columns, only primary and foreign keys or no columns)}

\item{focus}{A list of parameters for rendering (table filter)}

\item{col_attr}{Column atributes to display.
Only column name (\code{column}) is included by default.}
}
\description{
Create a graph object from data model object
}

