% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm.R
\name{dm_create_references}
\alias{dm_create_references}
\title{Create reference info}
\usage{
dm_create_references(col_table)
}
\arguments{
\item{col_table}{A data frame with table columns}
}
\description{
Creates references (foreign keys) based on reference table names in
column info table.
}
\details{
The function is used when creating data model object.
  \code{col_table} must have at least table, column and ref elements.
  When referencing to tables with compound primary keys
  additional ref_col with primary key columns must be provided.
}
\keyword{internal}

