% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{dm_export_graph}
\alias{dm_export_graph}
\title{Export graph to file}
\usage{
dm_export_graph(graph, file_name = NULL, file_type = NULL, width = NULL,
  height = NULL)
}
\arguments{
\item{graph}{a graph object}

\item{file_name}{file name}

\item{file_type}{file type (if not provided, file name extension is used)}

\item{width}{width}

\item{height}{height}
}
\description{
Export data model graph object object to PNG, PDF, PS or SVG file.
}
