% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm.R
\name{dm_from_data_frames}
\alias{dm_from_data_frames}
\title{Create data model object from R data frames}
\usage{
dm_from_data_frames(...)
}
\arguments{
\item{...}{Data frames or one list of data frames}
}
\description{
Uses data frame column names to create a data model diagram
}
