% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{dm_get_graph_svg}
\alias{dm_get_graph_svg}
\title{Get graph SVG}
\usage{
dm_get_graph_svg(graph)
}
\arguments{
\item{graph}{a graph object}
}
\value{
character in SVG format
}
\description{
Convert diagram graph object to SVG format
}
