% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dm.R
\name{dm_graph}
\alias{dm_graph}
\title{Create a graph object from data model object}
\usage{
dm_graph(x, rankdir = "BT", graph_name = "Data Model", graph_attrs = "",
  use_ref_ports = FALSE)
}
\arguments{
\item{x}{A data model object}

\item{rankdir}{Graph attribute for direction (eg. 'BT' = bottom --> top)}

\item{graph_name}{A graph name}

\item{graph_attrs}{Additional graph attributes}

\item{use_ref_ports}{If TRUE the outgoing links are positioned from
specific table elements (foreign keys). By default links are positioned
from table to table.}
}
\description{
Create a graph object from data model object
}

