% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm.R
\name{dm_list2coltable}
\alias{dm_list2coltable}
\title{List to column table}
\usage{
dm_list2coltable(x)
}
\arguments{
\item{x}{a named list}
}
\description{
Convert a 3 level named list to a data frame with column info
}
\details{
The funcion is used when creating data model object
  from list provided by yaml parser.
}
\keyword{internal}
