% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{dm_palette}
\alias{dm_palette}
\alias{dm_color_scheme}
\alias{dm_add_colors}
\alias{dm_get_color_scheme}
\alias{dm_set_color_scheme}
\title{Datamodel color schema}
\usage{
dm_palette(line_color = NULL, header_bgcolor, header_font, bgcolor)

dm_color_scheme(...)

dm_add_colors(color_scheme)

dm_get_color_scheme()

dm_set_color_scheme(color_scheme)
}
\arguments{
\item{line_color}{Rectangle color}

\item{header_bgcolor}{Table header background color}

\item{header_font}{Table header font color}

\item{bgcolor}{Table background color}

\item{...}{Palettes for color schema}

\item{color_scheme}{New colors created with dm_color_scheme}
}
\description{
Manage color schema for data model diagrams
}
\examples{
col_scheme <-
  dm_color_scheme(
    dm_palette(
      line_color = "#787878",
      header_bgcolor = "#A5A5A5",
      header_font = "#FFFFFF",
      bgcolor = "#E4E4E4"
    ),
    dm_palette(
      line_color = "#41719C",
      header_bgcolor = "#5B9BD5",
      header_font = "#FFFFFF",
      bgcolor = "#D6E1F1"
    ),
    dm_palette(
      line_color = "#BC8C00",
      header_bgcolor = "#FFC000",
      header_font = "#FFFFFF",
      bgcolor = "#FFEAD0"
    )
  )
}
\keyword{internal}
