% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm.R
\name{dm_re_query}
\alias{dm_re_query}
\title{Reverse engineer query}
\usage{
dm_re_query(rdbms)
}
\arguments{
\item{rdbms}{Which database ("postgres" or "sqlserver")}
}
\value{
A character string with sql query
}
\description{
Returns a string with SQL query to reverse engineer a database
}
\examples{
\dontrun{
library(RPostgreSQL)
# dvdrental sample database: http://www.postgresqltutorial.com/postgresql-sample-database
con <- dbConnect(dbDriver("PostgreSQL"), dbname="dvdrental", user ="postgres")
sQuery <- dm_re_query("postgres")
dm_dvdrental <- dbGetQuery(con, sQuery)
dbDisconnect(con)
}
}

