% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dm.R
\name{dm_read_yaml}
\alias{dm_read_yaml}
\title{Read YAML}
\usage{
dm_read_yaml(file = NULL, text = NULL)
}
\arguments{
\item{file}{A file in YAML format}

\item{text}{A YAML formated character string}
}
\description{
Reads a file in YAML format and returns a data model object.
}
\details{
YAML description should include table names (first level),
columns (second level) and column attributes (third level).
Expected (but not required) column attributes are
  \code{key} (Yes|No),
  \code{ref} (Name of referenced table),
  \code{comment} (column description).
}
\examples{
dm <-
  dm_read_yaml(text = "

    Person:
      Person ID: {key: yes}
      Name:
      E-mail:
      Street:
      Street number:
      City:
      ZIP:

    Order:
      Order ID: {key: yes}
      Customer: {ref: Person}
      Sales person: {ref: Person}
      Order date:
      Requested ship date:
      Status:

    Order Line:
      Order ID: {key: yes, ref: Order}
      Line number: {key: yes}
      Order item: {ref: Item}
      Quantity:
      Price:

    Item:
      Item ID: {key: yes}
      Item Name:
      Description:
  ")
}

