% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{dm_render_graph}
\alias{dm_render_graph}
\title{Render graph}
\usage{
dm_render_graph(graph, width = NULL, height = NULL)
}
\arguments{
\item{graph}{a graph object}

\item{width}{an optional parameter for specifying the width of the resulting
graphic in pixels.}

\item{height}{an optional parameter for specifying the height of the resulting graphic in pixels.}
}
\description{
Using DiagrammeR to render datamodelr graph object
}

