% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm.R
\name{dm_set_col_attr}
\alias{dm_set_col_attr}
\title{Set column attribute}
\usage{
dm_set_col_attr(dm, table, column, attr, value)
}
\arguments{
\item{dm}{A data model object}

\item{table}{Table name}

\item{column}{Column(s) name}

\item{attr}{Column attribute name}

\item{value}{New value}
}
\description{
Set column attribute value
}
\keyword{internal}

