% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm.R
\name{dm_set_display}
\alias{dm_set_display}
\title{Set table display}
\usage{
dm_set_display(dm, display)
}
\arguments{
\item{dm}{A data model object}

\item{display}{A named list of vectors with display as element names
and tables as values in vectors}
}
\description{
Change tables' display in a data model
}

