% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm.R
\name{dm_set_segment}
\alias{dm_set_segment}
\title{Set table segment}
\usage{
dm_set_segment(dm, table_segments)
}
\arguments{
\item{dm}{A data model object}

\item{table_segments}{A named list of vectors with segments as element names
and tables as values in vectors}
}
\description{
Change tables' segment name in a data model
}

