% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/graph.R
\name{dot_html_label}
\alias{dot_html_label}
\title{Create a label}
\usage{
dot_html_label(x, title, title_bgcolor = "#EFEBDD")
}
\arguments{
\item{x}{A data frame with column info}

\item{title}{A node title}

\item{title_bgcolor}{Title background color}
}
\description{
Create a label for dot HTML shape
}
\references{
See \url{http://www.graphviz.org/content/node-shapes}
}
\keyword{internal}

