% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm.R
\name{is.data_model}
\alias{is.data_model}
\title{Check if object is a data model}
\usage{
is.data_model(x)
}
\arguments{
\item{x}{Object to check if it is a data model}
}
\description{
Check if object is a data model
}
\keyword{internal}

