% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm.R
\name{print.data_model}
\alias{print.data_model}
\title{Print data model graph}
\usage{
\method{print}{data_model}(x, ...)
}
\arguments{
\item{x}{data model object.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print data model graph
}
