% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{print.datamodelr_graph}
\alias{print.datamodelr_graph}
\title{Print data model graph}
\usage{
\method{print}{datamodelr_graph}(x, ...)
}
\arguments{
\item{x}{data model object}

\item{...}{parameter passed to \link{dm_render_graph}}
}
\description{
Print data model graph
}
