% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_sql.R
\name{render_pg_query}
\alias{render_pg_query}
\title{Postgres ER Diagram Query}
\usage{
render_pg_query(schema)
}
\value{
Postgres SQL query as a string.
}
\description{
Returns a string with SQL query to reverse engineer a database
}
\examples{
\dontrun{
library(RPostgreSQL)
# dvdrental sample database: http://www.postgresqltutorial.com/postgresql-sample-database
con <- dbConnect(dbDriver("PostgreSQL"), dbname="dvdrental", user ="postgres")
sQuery <- dm_re_query("postgres")
dm_dvdrental <- dbGetQuery(con, sQuery)
dbDisconnect(con)
}
}
