% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{set_key}
\alias{set_key}
\title{Set key}
\usage{
set_key(dm, table, column)
}
\arguments{
\item{dm}{A data model object}

\item{table}{Table name}

\item{column}{Column(s) name}
}
\description{
Set column as a primary key
}
