% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{to_html_table}
\alias{to_html_table}
\title{Data frame to html table}
\usage{
to_html_table(x, title = "Table", attr_table, attr_header, attr_font,
  attr_td = NULL, trans = NULL, cols = names(x))
}
\arguments{
\item{x}{A data frame}

\item{attr_table}{A named list with table attributes}

\item{attr_header}{A named list with cell attributes in header cell}

\item{attr_td}{A function with parameters column, current row, value and
returns a named list with current cell attributes}

\item{trans}{Value transformation funcion}

\item{cols}{A named vector of columns to include in a table}
}
\description{
Used to create graphwiz dot HTML table labels
}
\keyword{internal}

