% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm.R
\name{as.data_model}
\alias{as.data_model}
\title{Coerce to a data model}
\usage{
as.data_model(x)
}
\arguments{
\item{x}{A data frame to be coerced to data model object}
}
\value{
If possible it returns a data model object.
  It is a list of data frames with at least the following columns:
  \item{ table }{A table name}
  \item{ column }{A name of the column in a table}
  \item{ key }{A boolean value indicating this column is in a primary key.
  Use integer values with order for compound keys }
  \item{ ref }{A character string with a referenced table name.
    If exists (not NA) then the column is a foreign key.}
  \item{ ref_col }{A character string with a referenced table name column.
    This is only necessary when referenced table has multi-column key}
}
\description{
Functions to coerce an object to a data model if possible.
}
\details{
Function accepts a data frame with columns info.
  Data frame must have 'table' and 'column' elements.
  Optional element 'key' (boolean) marks a column as a primary key.
  Optional element 'ref' (character string) defines referenced table name.
  Optional element 'ref_col' (character string) defines a column in a
    referenced table name primary key (only necessery when referenced
    table has a compound primary key).
}
