% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dm.R
\docType{package}
\name{datamodelr-package}
\alias{datamodelr}
\alias{datamodelr-package}
\title{datamodelr: data model diagrams}
\description{
Provides a simple structure to describe data models,
functions to read data model from YAML file,
and a function to create DiagrammeR graph objects:
}
\details{
\itemize{
  \item \pkg{datamodelr}'s data model object is a simple list of data frames which
    represent the model entities and include elements and their relations.
    See \code{\link{as.data_model}}.
  \item Function \code{\link{dm_read_yaml}} reads YAML format and creates a
    data model object.
  \item Function \code{\link{dm_graph}} creates DiagrammeR graph object from
    data model object.
}
}

