% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dm.R
\name{dm_re_query}
\alias{dm_re_query}
\title{Reverse engineer query}
\usage{
dm_re_query(rdbms)
}
\arguments{
\item{rdbms}{Which database ("postgres" or "sqlserver" for now)}
}
\value{
A character string with sql query
}
\description{
Returns a query string to reverse engineer a database
}

