% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{dm_render_graph}
\alias{dm_render_graph}
\title{Render graph}
\usage{
dm_render_graph(graph, output = "graph", layout = NULL, width = NULL,
  height = NULL)
}
\arguments{
\item{graph}{a \pkg{DiagrammeR} dgr_graph object}

\item{output}{string specifying the output type; graph (the default) renders
the graph using the grViz function,
vivagraph renders the graph using the vivagraph function,
visNetwork renders the graph using the visnetwork function,
DOT outputs DOT code for the graph, and SVG provides SVG code for the rendered graph.}

\item{layout}{a string specifying a layout type for a vivagraph rendering of the graph,
either forceDirected or constant.}

\item{width}{an optional parameter for specifying the width of the resulting
graphic in pixels.}

\item{height}{an optional parameter for specifying the height of the resulting graphic in pixels.}
}
\description{
A wrapper around DiagrammeR::render_graph
}

