% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/graph.R
\name{dot_html_label}
\alias{dot_html_label}
\title{Create a label}
\usage{
dot_html_label(x, title, palette_id = "default", col_attr = c("column"))
}
\arguments{
\item{x}{A data frame with column info}

\item{title}{A node title}

\item{palette_id}{Which color palette should be used (default)}

\item{col_attr}{Column atributes to display.
Only column name (\code{column}) is included by default.}
}
\description{
Create a label for dot HTML shape
}
\references{
See \url{http://www.graphviz.org/content/node-shapes}
}
\keyword{internal}

